added Gates into method

+++ laravel-8-todo/app/Http/Controllers/TodoController.php
@@ -3,31 +3,40 @@
 namespace App\Http\Controllers;
 
 use App\Models\Todo;
+
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Redirect;
 
 class TodoController extends Controller
 {
 
+
+    public function __construct()
+    {
+        $this->middleware('auth');
+    }
+
     public function index()
     {
-        $todos = Todo::latest()->paginate(3);
+
+        $todos = auth()->user()->todos()->paginate(3);
+
         return view('todos.index', compact('todos'));
     }
 
-
     public function show(Todo $todo)
     {
+        Gate::authorize('view', $todo);
 
         return view('todos.show', compact('todo'));
+
     }
 
     public function create()
     {
-
         return view('todos.create');
     }
-
 
     public function store(Request $request)
     {
@@ -39,7 +48,8 @@ class TodoController extends Controller
 
         Todo::create([
             'title' => $request->title,
-            'description' => $request->description
+            'description' => $request->description,
+            'user_id' => auth()->user()->id
         ]);
 
         alert()->success('تسک با موفقیت انجام شد.', 'با تشکر');
@@ -49,18 +59,23 @@ class TodoController extends Controller
 
     public function edit(Todo $todo)
     {
+
+        Gate::authorize('update', $todo);
+
         return view('todos.edit', compact('todo'));
 
     }
 
-
     public function update(Request $request, Todo $todo)
     {
+        Gate::authorize('update', $todo);
+
 
         $request->validate([
             'title' => 'required',
             'description' => 'required'
         ]);
+
 
         Todo::updated([
             'title' => $request->title,
@@ -74,6 +89,8 @@ class TodoController extends Controller
 
     public function destroy(Todo $todo)
     {
+        Gate::authorize('delete', $todo);
+
         $todo->delete();
         alert()->error('تسک با موفقیت حذف شد.', 'دقت کنید');
         return redirect()->route('todos.index');
@@ -82,11 +99,13 @@ class TodoController extends Controller
 
     public function complete(Todo $todo)
     {
+        Gate::authorize('complete', $todo);
+
         $todo->update([
             'completed' => 1
         ]);
 
-        alert()->success('تسک مورد نظر به وضعیت انجام شد تغییر پیدا کرد.','با تشکر');
+        alert()->success('تسک مورد نظر به وضعیت انجام شد تغییر پیدا کرد.', 'با تشکر');
         return redirect()->route('todos.index');
 
     }
